move some .set to initialization, shorttened some code, changed tokencache to an interface with get, set, remove
