+++ tests/run_test.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"encoding/json"
 	"testing"
+	"time"
 
 	loafer "github.com/arkjxu/loafer"
 )
@@ -33,22 +34,36 @@ func handleDevCommand(ctx *loafer.SlackContext) {
 	})
 }
 
-func tokenCache(_ string) []loafer.SlackAuthToken {
-	tokens := []loafer.SlackAuthToken{}
-	return tokens
+type TokenCache struct {
+	tokens map[string]string
+}
+
+func (t *TokenCache) Get(workspace string) string {
+	return t.tokens[workspace]
+}
+
+func (t *TokenCache) Set(workspace string, token string) {
+	t.tokens[workspace] = token
+}
+
+func (t *TokenCache) Remove(workspace string) {
+	delete(t.tokens, workspace)
 }
 
 func TestRun(t *testing.T) {
+	myTokenCache := TokenCache{
+		tokens: make(map[string]string)}
 	opts := loafer.SlackAppOptions{
 		Name:          "Dev Bot",
 		Prefix:        "dev",
-		TokensCache:   tokenCache,
+		TokensCache:   &myTokenCache,
 		SigningSecret: "xxxxxxxxxxxxxxxxxxxxxxxxxxxxx",
 		ClientID:      "xxxxxxxxxxxx.xxxxxxxxxx",
 		ClientSecret:  "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"}
 	app := loafer.InitializeSlackApp(&opts)
 	app.OnCommand("/coaching", handleDevCommand)
 	app.ServeApp(8080, func() {
+		time.Sleep(3 * time.Second)
 		timeOut, cancel := context.WithTimeout(context.Background(), 1000)
 		defer cancel()
 		app.Close(timeOut)
